moved NewsForm functions to NewsMain and passed functions and state  as props to NewsForm. state is now being updated in NewsMain and rerendering when a new NewsItem is created. Now getting warning that a key is needed--need to fix this.
